feat(init): refactor template files
