+++ lib/executeInstall.js
@@ -11,21 +11,15 @@ const devInstall =
   'npm i -D eslint-config-airbnb-base eslint nodemon supertest jest @babel/cli @babel/core @babel/node @babel/preset-env';
 const depInstall = 'npm i express body-parser cors dotenv';
 
-const executeInstall = () => {
-  // install dev dependencies
-  runner({ command: devInstall, spinner: devSpinner }).then(() => {
+const executeInstall = (root) => {
+ //install dev dependencies
+  runner({ command: `cd ${root} && ${devInstall}`, spinner: devSpinner }).then(() => {
     // then install dependencies
     runner({
-      command: depInstall,
+      command: `cd ${root} && ${depInstall}`,
       spinner: depSpinner,
       onComplete: packageComplete,
-    }).then(() => {
-      // then start dev server
-      runner({
-        command: 'npm run dev',
-        spinner: serverSpinner,
-      });
-    });
+    })
   });
 };
