+++ lib/generateFiles.js
@@ -1,5 +1,5 @@
 const fs = require('fs');
-const helpers = require('./template');
+const helpers = require('../templates');
 const execInstall = require('./executeInstall');
 const { checkWriteFile, writeFile } = require('./utils');
 
@@ -8,6 +8,10 @@ const errorhandler = err => {
 };
 
 const generateFiles = project => {
+  const rootFolderName = project.name
+    .toLowerCase()
+    .split(' ')
+    .join('-');
   const packageJson = helpers.generatePackageJson(project);
   const eslint = helpers.generateEslint();
   const babel = helpers.generateBabel();
@@ -25,71 +29,76 @@ const generateFiles = project => {
   const v1Index = helpers.generateV1Index();
 
   const filesObj = [
-    { path: '.gitignore', data: gitignore },
-    { path: '.babelrc', data: babel },
-    { path: '.env', data: dotenv },
-    { path: '.env.sample', data: '' },
-    { path: 'jest.config.js', data: jestConfig },
-    { path: '.eslintrc.js', data: eslint },
-    { path: 'package.json', data: packageJson, sync: true },
+    { path: rootFolderName, name: '.gitignore', data: gitignore },
+    { path: rootFolderName, name: '.babelrc', data: babel },
+    { path: rootFolderName, name: '.env', data: dotenv },
+    { path: rootFolderName, name: '.env.sample', data: '' },
+    { path: rootFolderName, name: 'jest.config.js', data: jestConfig },
+    { path: rootFolderName, name: '.eslintrc.js', data: eslint },
     {
-      path: `${project.root}/tests`,
+      path: rootFolderName,
+      name: 'package.json',
+      data: packageJson,
+      sync: true,
+    },
+    {
+      path: `${rootFolderName}/${project.root}/tests`,
       name: 'welcome.spec.js',
       data: welcomeTest,
     },
     {
-        path: `${project.root}/tests`,
-        name: 'sample.spec.js',
-        data: sampleTest,
-      },
+      path: `${rootFolderName}/${project.root}/tests`,
+      name: 'sample.spec.js',
+      data: sampleTest,
+    },
     {
-      path: `${project.root}/config`,
+      path: `${rootFolderName}/${project.root}/config`,
       name: 'envVariables.js',
       data: envConfig,
       sync: true,
       check: true,
     },
     {
-      path: `${project.root}/controllers/sample`,
+      path: `${rootFolderName}/${project.root}/controllers/sample`,
       name: 'index.js',
       data: sampleController,
       sync: true,
       check: true,
     },
     {
-      path: `${project.root}/routes/v1`,
+      path: `${rootFolderName}/${project.root}/routes/v1`,
       name: 'sample.js',
       data: sampleRoute,
       sync: true,
     },
     {
-      path: `${project.root}/routes/v1`,
+      path: `${rootFolderName}/${project.root}/routes/v1`,
       name: 'index.js',
       data: v1Index,
       sync: true,
     },
     {
-      path: `${project.root}/routes`,
+      path: `${rootFolderName}/${project.root}/routes`,
       name: 'index.js',
       data: rootRoute,
       sync: true,
     },
     {
-      path: `${project.root}`,
+      path: `${rootFolderName}/${project.root}`,
       name: 'server.js',
       data: server,
       sync: true,
     },
     {
-      path: `${project.root}`,
+      path: `${rootFolderName}/${project.root}`,
       name: 'index.js',
       data: serverIndex,
       sync: true,
     },
   ];
- 
+
   fs.mkdir(
-    `${project.root}/middlewares/validators`,
+    `${rootFolderName}/${project.root}/middlewares/validators`,
     { recursive: true },
     errorhandler
   );
@@ -101,8 +110,8 @@ const generateFiles = project => {
       writeFile(path, data, sync || false);
     }
   });
-
- execInstall()
+ 
+  execInstall(rootFolderName)
 };
 
 module.exports = generateFiles;
