+++ lib/init.js
@@ -1,10 +1,11 @@
 const { getProjectInfo, intro } = require('./utils');
 const generateFiles = require('./generateFiles');
 
-const init = () => {
+const init = name => {
   intro();
-  getProjectInfo().then(data => {
-    generateFiles(data);
+  getProjectInfo(name).then(data => {
+    const dataObj = { name, ...data };
+    generateFiles(dataObj);
   });
 };
