+++ lib/utils.js
@@ -24,15 +24,6 @@ const intro = () => {
   );
 };
 
-const customServerMsg = () =>
-  console.log(
-    chalk.greenBright(
-      `\n✔ Server is running on port 4000. Go to ${chalk.blue.bold.underline(
-        'http://localhost:4000/'
-      )}`
-    )
-  );
-
 const runner = ({ command, onComplete, spinner }) => {
   spinner.start();
 
@@ -43,10 +34,7 @@ const runner = ({ command, onComplete, spinner }) => {
     //listen to every data stream
     newProcess.stdout.on('data', data => {
       spinner.stop(); // stop spinner before printing output
-      // check if data stream output is listening to server print custom message
-      if (data.toString().trim() === 'Server is running on port 4000') {
-       customServerMsg();
-      } else console.log(`${data}`);
+      console.log(`${data}`);
     });
 
     // on exit listener
@@ -57,7 +45,7 @@ const runner = ({ command, onComplete, spinner }) => {
   });
 };
 
-const getProjectInfo = () => {
+const getProjectInfo = name => {
   const questions = [
     {
       type: 'input',
@@ -90,6 +78,7 @@ const getProjectInfo = () => {
       },
     },
   ];
+  if (name) questions.splice(0, 1);
   return inquirer.prompt(questions);
 };
 
@@ -122,5 +111,5 @@ module.exports = {
   runner,
   getProjectInfo,
   intro,
-  packageComplete
+  packageComplete,
 };
