+++ templates/generateBabel.js
@@ -1,0 +1,12 @@
+module.exports = () => `{
+    "presets": [
+      [
+        "@babel/preset-env",
+        {
+          "targets": {
+            "node": "current"
+          }
+        }
+      ]
+    ]
+}`
