+++ templates/generateEnvConfig.js
@@ -1,0 +1,5 @@
+module.exports = () => `import dotenv from 'dotenv';
+dotenv.config();
+
+export const  { PORT, NODE_ENV } = process.env;
+`
