+++ templates/generateGitIgnore.js
@@ -1,0 +1,6 @@
+module.exports = () => `
+node_modules
+*.env
+coverage
+dist
+`
