+++ templates/generateJestConfig.js
@@ -1,0 +1,5 @@
+module.exports = () => `module.exports = {
+  testPathIgnorePatterns: ['/node_modules/'],
+  testEnvironment: 'node',
+};
+`
