+++ templates/generateRootRoute.js
@@ -1,0 +1,6 @@
+module.exports = () => `import v1 from './v1';
+
+export default (app) => {
+  app.use('/api/v1', v1);
+};
+`
