+++ templates/generateSampleTest.js
@@ -1,0 +1,15 @@
+module.exports = (name) => `import supertest from 'supertest';
+import { app } from '../server.js';
+
+const server = supertest(app);
+
+describe('GET Sample message /', () => {
+  test('should get a sample message from sample rout', async done => {
+    const res = await server.get('/api/v1/sample');
+  
+    expect(res.status).toBe(200);
+    expect(res.body.message).toEqual('This is a sample route');
+    done();
+  });
+});
+`
